Repost impemented

+++ app/Controllers/Http/TweetController.js
@@ -1,26 +1,39 @@
 'use strict'
+const { validate } = use('Validator')
 
 const Tweet = use('App/Models/Tweet');
 const User = use('App/Models/User');
 
+const TweetRules = {
+  title: 'required|max200',
+  content: 'required|max:1000'
+}
+
 class TweetController {
+
     async create ({ request, response, auth}) {
         const data = request.all();
 
-        let tweet = new Tweet();
-        tweet.title = data.title;        
-        tweet.content = data.content;     
+        const validation = await validate(request.all(), TweetRules)
 
-        try {
-            const user = await auth.getUser();
+        if (validation.fails()) {
+            return validation.messages();
+        } else { 
+            let tweet = new Tweet();
+            tweet.title = data.title;        
+            tweet.content = data.content;     
+    
+            try {
+                const user = await auth.getUser();
                 await user.tweets().create({
                     title : data.title,
                     content : data.content
                 });
-        } catch (error) {
-            response.send("error");
+                response.send("Tweeted Successfully.")
+            } catch (error) {
+                response.send("error");
+            }
         }
-        
     }
 
     async read ({ request, response, auth, params}) {
@@ -38,11 +51,12 @@ class TweetController {
         let id = request.all().id;
         try {
             const user = await auth.getUser();
-            try {
+            const exists = await user.tweets().where('id','=',id);
+            if(exists.row.length == 0) {
+                response.send('You do not own the tweet.');
+            } else {
                 await user.tweets().where('id','=',id).delete();
                 response.send('Deleted Successfully.');
-            } catch (error) {
-                response.send(error);
             }
         } catch (error) {
             response.send(error);
@@ -54,8 +68,12 @@ class TweetController {
         try {
             const user = await auth.getUser();
             let tweet = await Tweet.findBy('id',id);
-            await tweet.likes().attach(user.id);
-            response.send('Liked Successfully.');
+            if("id" in tweet) {
+                await tweet.likes().attach(user.id);
+                response.send('Liked Successfully.');
+            } else {
+                response.send('Tweet not found.');                
+            }
         } catch (error) {
             response.send(error);
         }
@@ -66,8 +84,12 @@ class TweetController {
         try {
             const user = await auth.getUser();
             let tweet = await Tweet.findBy('id',id);
-            await tweet.likes().detach(user.id);
-            response.send('Un-liked Successfully.');
+            if("id" in tweet) {
+                await tweet.likes().detach(user.id);
+                response.send('Un-liked Successfully.');
+            } else {
+                response.send('Tweet not found.');                
+            }
         } catch (error) {
             response.send(error);
         }
@@ -85,6 +107,25 @@ class TweetController {
 
     async getAllTweets ({ request, response, auth, params}) { 
         return await Tweet.all();
+    }
+
+    async repost ({ request, response, auth, params}) { 
+        let id = request.all().id;
+        try {
+            const user = await auth.getUser();
+            let tweet = await Tweet.findBy('id',id);
+            if("id" in tweet) {
+                await user.tweets().create({
+                    title : tweet.title,
+                    content : tweet.content
+                });
+                response.send('Re-posted Successfully.');
+            } else {
+                response.send('Tweet not found.');                
+            }
+        } catch (error) {
+            response.send(error);
+        }
     }
 }
